Reorganized project folders
