+++ src/main/java/org/sakaiproject/elfinder/controller/executors/SakaiUploadCommandExecutor.java
@@ -1,4 +1,4 @@
-package org.sakaiproject.elfinder;
+package org.sakaiproject.elfinder.controller.executors;
 
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -11,6 +11,7 @@ import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.fileupload.FileItemStream;
 import org.apache.commons.io.IOUtils;
 import org.json.JSONObject;
+import org.sakaiproject.elfinder.impl.SakaiFsService;
 
 import cn.bluejoe.elfinder.controller.executor.AbstractJsonCommandExecutor;
 import cn.bluejoe.elfinder.controller.executor.CommandExecutor;
