+++ src/main/java/org/sakaiproject/elfinder/impl/SakaiFsServiceFactory.java
@@ -1,4 +1,4 @@
-package org.sakaiproject.elfinder;
+package org.sakaiproject.elfinder.impl;
 
 import cn.bluejoe.elfinder.service.FsService;
 import cn.bluejoe.elfinder.service.FsServiceFactory;
